Move store to own file.
